--- src/main/kotlin/lotto/controller/LottoOrderDataHttpRequest.kt
@@ -1,6 +1,0 @@
-package lotto.controller
-
-data class LottoOrderDataHttpRequest(
-    val number: LottoRequest
-){
-}
